# [Sofa.Type] Speed up fixed_array constructors with perfect forwarding